Refactored codebase to use reuseable component
