+++ Jobhunt365/src/Components/ReuseableComponets/JobCard/JobCard.tsx
@@ -19,9 +19,9 @@ const JobCard = ({ JobStatus }: { JobStatus: string }) => {
         <div className='flex items-center gap-4 w-full'>
           <LogoComponent company='xoogle' imageUrl={google} />
           <div className='flex flex-col '>
-            <h2 className='font-bold text-[#4b5563]  max-w-[150px] truncate'>
+            <h5 className='font-bold text-[#4b5563]  max-w-[150px] truncate'>
               Senior Frontend Developer
-            </h2>
+            </h5>
             <p className='text-sm text-[#94a3b8]'>Goldman Sach</p>
           </div>
         </div>
